LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-area-scrolling-003.tentative-expected.txt



FAIL Initial scroll position assert_equals: e1 x expected 100 but got 0
FAIL Scroll to 40,60 assert_equals: e1 x expected 100 but got 0
FAIL Reattach at 40,60 assert_equals: e1 x expected 100 but got 0
FAIL Redisplay at 40,60 assert_equals: e1 x expected 120 but got 0
FAIL Scroll to 0,0 and relayout assert_equals: e1 x expected 120 but got 0
FAIL Redisplay at 0,0 assert_equals: e1 x expected 100 but got 0
